fix global install
I haven't tested this, but I believe that this is what would need to happen to make it so global installs work. Right now if you install this package globally, it'll try to resolve plugins where `eslint-to-editorconfig` is located. Because most people install plugins locally though this isn't usable this way. By using `process.cwd()` it makes everything relative to where the command is executed.